refactor: update photo entity configuration and queries to use new constants
- Replace references to old constants with new photo configuration
- Introduce new constants for photo and collection defaults
- Update queries and components to align with the new structure
