+++ apps/web/src/entities/photo/config/photo.ts
@@ -12,14 +12,14 @@ export const ALBUM_OPTIONS = {
 	},
 } as const;
 
-export const TRENDING_PHOTO_PARAMS = {
-	LIMIT: 8,
+export const TRENDING_PHOTO_DEFAULTS = {
+	LIMIT: 12,
 	QUERY: '',
 	SORT: 'popular',
 } as const;
 
-export const PHOTO_PARAMS = {
-	LIMIT: 8,
+export const PHOTO_DEFAULTS = {
+	LIMIT: 12,
 	QUERY: '',
 	SORT: 'latest',
 } as const;
@@ -30,6 +30,6 @@ export const PRIORITY_IMAGE = {
 	DESKTOP: 6,
 } as const;
 
-export const COLLECTION_PARAMS = {
+export const COLLECTION_DEFAULTS = {
 	sort: 'latest',
 } as const;
