+++ apps/web/src/entities/photo/index.ts
@@ -1,5 +1,5 @@
 export * from './api';
-export * from './config/constants';
+export * from './config/photo';
 export { transformPhoto } from './lib/transformPhoto';
 export { usePhotoFilterStore } from './model/usePhotoFilterStore';
 export * from './ui';
